+++ packages/atomic/src/components/commerce/atomic-product-field-condition/atomic-product-field-condition.new.stories.tsx
@@ -20,7 +20,7 @@ const {decorator: productTemplateDecorator} = wrapInProductTemplate();
 
 const meta: Meta = {
   component: 'atomic-product-field-condition',
-  title: 'Commerce/atomic-product-field-condition',
+  title: 'Commerce/Product Field Condition',
   id: 'atomic-product-field-condition',
   render: renderComponent,
   decorators: [