+++ packages/atomic/src/components/commerce/atomic-product-rating/atomic-product-rating.new.stories.tsx
@@ -32,7 +32,7 @@ const {decorator: productTemplateDecorator} = wrapInProductTemplate();
 
 const meta: Meta = {
   component: 'atomic-product-rating',
-  title: 'Commerce/atomic-product-rating',
+  title: 'Commerce/Product Rating',
   id: 'atomic-product-rating',
   render: renderComponent,
   decorators: [