+++ packages/atomic/src/components/commerce/atomic-product-rating/atomic-product-rating.ts
@@ -24,8 +24,6 @@ import type {CommerceBindings} from '../atomic-commerce-interface/atomic-commerc
  * @cssprop --atomic-rating-icon-active-color - Color of the icon when active.
  * @cssprop --atomic-rating-icon-inactive-color - Color of the icon when inactive.
  * @cssprop --atomic-rating-icon-outline - Outline color of the icon.
- *
- * @alpha
  */
 @customElement('atomic-product-rating')
 @bindings()