+++ packages/atomic/src/components/common/facets/facet-value-checkbox/facet-value-checkbox.ts
@@ -27,6 +27,7 @@ export const renderFacetValueCheckbox: FunctionalComponentWithChildren<
     const ariaLabelAttributes = {
       value: props.displayValue,
       count: props.numberOfResults,
+      formattedCount: count,
       interpolation: {escapeValue: false},
     };
     const selectedAriaLabel = props.i18n.t('facet-value', ariaLabelAttributes);