+++ packages/atomic/src/components/search/facets/atomic-timeframe-facet/atomic-timeframe-facet.new.stories.tsx
@@ -1,4 +1,8 @@
-import type {Meta, StoryObj as Story} from '@storybook/web-components';
+import type {
+  Decorator,
+  Meta,
+  StoryObj as Story,
+} from '@storybook/web-components';
 import {html} from 'lit';
 import {within} from 'shadow-dom-testing-library';
 import {parameters} from '@/storybook-utils/common/common-meta-parameters';
@@ -14,6 +18,9 @@ const {decorator, play} = wrapInSearchInterface({
     return r;
   },
 });
+
+const commerceFacetWidthDecorator: Decorator = (story) =>
+  html`<div style="min-width: 470px;">${story()}</div> `;
 
 const meta: Meta = {
   component: 'atomic-timeframe-facet',
@@ -40,7 +47,7 @@ const meta: Meta = {
   `,
   },
   render: renderComponent,
-  decorators: [decorator],
+  decorators: [commerceFacetWidthDecorator, decorator],
   parameters,
   play,
 };